+++ hal/internet/youtube.py
@@ -22,8 +22,11 @@ from bs4 import BeautifulSoup
 
 from hal.internet.web import Webpage
 
+YOUTUBE_USER_BASEURL = "https://www.youtube.com/user/"
+YOUTUBE_FEED_BASEURL = "https://www.youtube.com/feeds/videos.xml?channel_id="
 
-def get_channel_page(channel_name, youtube_channel_url="https://www.youtube.com/user/"):
+
+def get_channel_page_from_name(channel_name):
     """
     :param channel_name: string
         name of channel (e.g in "https://www.youtube.com/user/caseyneistat" you should take "caseyneistat")
@@ -33,37 +36,59 @@ def get_channel_page(channel_name, youtube_channel_url="https://www.youtube.com/
         source page of youtube channel.
     """
 
-    channel_url = youtube_channel_url + channel_name  # url of channel
+    channel_url = YOUTUBE_USER_BASEURL + channel_name  # url of channel
     source_page = Webpage(channel_url).get_html_source()  # get source page of channel homepage
     return source_page
 
 
-def get_channel_id(channel_name, channel_id_field="data-channel-external-id"):
+def get_channel_id_from_name(channel_name):
     """
     :param channel_name: string
         name of channel (e.g in "https://www.youtube.com/user/caseyneistat" you should take "caseyneistat")
-    :param channel_id_field: string
-        default field to get channel id
     :return string
         id of youtube channel
     """
 
-    soup = BeautifulSoup(get_channel_page(channel_name), "lxml")  # parser for source page
+    soup = BeautifulSoup(get_channel_page_from_name(channel_name), "lxml")  # parser for source page
     channel_id = soup.find_all("span", {"class": "channel-header-subscription-button-container"})  # get all good spans
     channel_id = channel_id[0].find_all("button")[0]  # get button in first span
-    channel_id = channel_id[channel_id_field]  # get id
+    channel_id = channel_id["data-channel-external-id"]  # get id
     return channel_id
 
 
-def get_channel_feed_url(channel_name, base_feed_url="https://www.youtube.com/feeds/videos.xml?channel_id="):
+def get_channel_feed_url_from_id(channel_id):
     """
-    :param channel_name: string
-        name of channel (e.g in "https://www.youtube.com/user/caseyneistat" you should take "caseyneistat")
-    :param base_feed_url: string
-        default base url for rss feed of youtube channels.
+    :param channel_id: string
+        Id of channel (e.g in "https://www.youtube.com/channel/UC2zjki3bJIaXmgV_LBQ2jTg" you should take "UC2zjki3bJIaXmgV_LBQ2jTg")
     :return string
         rss url feed of youtube channel.
     """
 
-    channel_id = get_channel_id(channel_name)  # get channel id
-    return base_feed_url + channel_id
+    return YOUTUBE_FEED_BASEURL + channel_id
+
+
+def get_channel_feed_url_from_name(channel_name):
+    """
+    :param channel_name: string
+        name of channel (e.g in "https://www.youtube.com/user/caseyneistat" you should take "caseyneistat")
+    :return string
+        rss url feed of youtube channel.
+    """
+
+    channel_id = get_channel_id_from_name(channel_name)  # get channel id
+    return get_channel_feed_url_from_id(channel_id)
+
+
+def get_channel_feed_url_from_video(video_url):
+    """
+    :param video_url: string
+        Url of video (e.g in https://www.youtube.com/watch?v=KB_iTbDrkxE)
+    :return string
+        rss url feed of youtube channel.
+    """
+
+    web_page = Webpage(video_url)
+    web_page.get_html_source()
+    channel_id = web_page.soup.find_all("div", {"class": "yt-user-info"})[0].a["href"]
+    channel_id = str(channel_id).strip().replace("/channel/", "")  # get channel id
+    return get_channel_feed_url_from_id(channel_id)